User Middleware - Milestone 3

+++ src/app/delete-product/delete-product.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MovieDAO } from '../service/MovieDAO';
+import { UserDAO } from '../service/UserDAO';
 
 @Component({
   selector: 'app-delete-product',
@@ -9,23 +10,26 @@ import { MovieDAO } from '../service/MovieDAO';
 })
 export class DeleteProductComponent implements OnInit {
 
-  constructor(private route: ActivatedRoute, private service: MovieDAO, private router: Router) { }
+  constructor(private route: ActivatedRoute, private middleware: UserDAO, private service: MovieDAO, private router: Router) { }
 
   id: number;
   deleted: boolean = false;
 
   ngOnInit(): void 
   {
-    this.id = Number(this.route.snapshot.paramMap.get('id'));
-    this.service.deleteMovie( this.id, (any) =>
+    if(this.middleware.currentUser==undefined)
     {
-      this.deleted = true;
-      alert("Movie has been deleted");
-        //console.log(product);
-        this.router.navigateByUrl('/');
-    });
-    
-
+      this.router.navigateByUrl('/login');
+    }
+    else {
+      this.id = Number(this.route.snapshot.paramMap.get('id'));
+      this.service.deleteMovie( this.id, (any) =>
+      {
+        this.deleted = true;
+        alert("Movie has been deleted");
+          //console.log(movie);
+          this.router.navigateByUrl('/');
+      });
+    }
   }
-
 }
